+++ Twilio/Rest/Api/V2010/Account/Call/RecordingInstance.php
@@ -26,11 +26,11 @@ use Twilio\Version;
  * @property string price
  * @property string uri
  * @property array encryptionDetails
- * @property integer errorCode
- * @property string status
- * @property string source
- * @property integer channels
  * @property string priceUnit
+ * @property string status
+ * @property integer channels
+ * @property string source
+ * @property integer errorCode
  */
 class RecordingInstance extends InstanceResource {
     /**
@@ -58,11 +58,11 @@ class RecordingInstance extends InstanceResource {
             'price' => Values::array_get($payload, 'price'),
             'uri' => Values::array_get($payload, 'uri'),
             'encryptionDetails' => Values::array_get($payload, 'encryption_details'),
-            'errorCode' => Values::array_get($payload, 'error_code'),
-            'status' => Values::array_get($payload, 'status'),
-            'source' => Values::array_get($payload, 'source'),
-            'channels' => Values::array_get($payload, 'channels'),
             'priceUnit' => Values::array_get($payload, 'price_unit'),
+            'status' => Values::array_get($payload, 'status'),
+            'channels' => Values::array_get($payload, 'channels'),
+            'source' => Values::array_get($payload, 'source'),
+            'errorCode' => Values::array_get($payload, 'error_code'),
         );
 
         $this->solution = array(