+++ Twilio/Tests/Integration/Notify/V1/ServiceTest.php
@@ -51,6 +51,7 @@ class ServiceTest extends HolodeckTestCase {
                 "default_fcm_notification_protocol_version": "3",
                 "default_alexa_notification_protocol_version": "3",
                 "log_enabled": true,
+                "type": "S",
                 "url": "https://notify.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
                     "bindings": "https://notify.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Bindings",
@@ -127,6 +128,7 @@ class ServiceTest extends HolodeckTestCase {
                 "default_fcm_notification_protocol_version": "3",
                 "default_alexa_notification_protocol_version": "3",
                 "log_enabled": true,
+                "type": "S",
                 "url": "https://notify.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
                     "bindings": "https://notify.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Bindings",
@@ -189,6 +191,7 @@ class ServiceTest extends HolodeckTestCase {
                         "default_fcm_notification_protocol_version": "3",
                         "default_alexa_notification_protocol_version": "3",
                         "log_enabled": true,
+                        "type": "S",
                         "url": "https://notify.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "links": {
                             "bindings": "https://notify.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Bindings",
@@ -266,6 +269,7 @@ class ServiceTest extends HolodeckTestCase {
                 "alexa_skill_id": null,
                 "facebook_messenger_page_id": "4",
                 "log_enabled": true,
+                "type": "S",
                 "url": "https://notify.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
                     "bindings": "https://notify.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Bindings",