+++ Twilio/Tests/Integration/Proxy/V1/Service/SessionTest.php
@@ -37,7 +37,7 @@ class SessionTest extends HolodeckTestCase {
             '
             {
                 "service_sid": "KSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "status": "in-progess",
+                "status": "in-progress",
                 "unique_name": "unique_name",
                 "date_started": "2015-07-30T20:00:00Z",
                 "date_ended": "2015-07-30T20:00:00Z",
@@ -126,7 +126,7 @@ class SessionTest extends HolodeckTestCase {
             '
             {
                 "service_sid": "KSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "status": "in-progess",
+                "status": "in-progress",
                 "unique_name": "unique_name",
                 "date_started": "2015-07-30T20:00:00Z",
                 "date_ended": "2015-07-30T20:00:00Z",
@@ -202,7 +202,7 @@ class SessionTest extends HolodeckTestCase {
             '
             {
                 "service_sid": "KSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "status": "in-progess",
+                "status": "in-progress",
                 "unique_name": "unique_name",
                 "date_started": "2015-07-30T20:00:00Z",
                 "date_ended": "2015-07-30T20:00:00Z",